refactor(davinci): split helper order buffer logic
